+++ src/Resources/contao/languages/de/tl_list_config.php
@@ -1,0 +1,11 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_list_config'];
+
+/**
+ * Fields
+ */
+$lang['addSlick'][0]    = 'Als Slick-Slider ausgeben';
+$lang['addSlick'][1]    = 'Wählen Sie diese Option, um die Liste als Slick-Slider zu rendern.';
+$lang['slickConfig'][0] = 'Slick-Konfiguration';
+$lang['slickConfig'][1] = 'Wählen Sie hier die gewünschte Slick-Konfiguration aus.';