+++ src/Resources/contao/languages/en/tl_list_config.php
@@ -1,0 +1,11 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_list_config'];
+
+/**
+ * Fields
+ */
+$lang['addSlick'][0]    = 'Render as slick slider';
+$lang['addSlick'][1]    = 'Choose this option to render the list as slick slider.';
+$lang['slickConfig'][0] = 'Slick Configuration';
+$lang['slickConfig'][1] = 'Select the configuration for slick slider.';