+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.2.0] - 2019-10-17
+- added list bundle support (integrated contao-slick-list-bundle)
+- replace contao-slick and contao-slick-list-bundle via composer.json
+
 
 ## [1.1.1] - 2019-10-16
 - fixed svg image files not working with content elements (#15)